requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
covid==2.4.1
cycler==0.10.0
discord.py==1.7.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
nasapy==0.2.7
newsapi-python==0.2.6
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pydantic==1.8.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pyttsx3==2.90
pytz==2021.1
requests==2.12.5
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tekore==3.7.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
wit==6.0.0
yarl==1.6.3